move FormattedNumber
